streamlit-app/requirements.txt
torch==2.8.0
torchvision==0.23.0
pillow==11.3.0
Streamlit==1.48.0
